sessions & orders & customers management

+++ app/services/employee.py
@@ -1,35 +1,13 @@
 from typing import List
-import uuid
 from sqlalchemy import func
 from sqlalchemy.orm import Session
 from app import models, schemas
-from app.OAuth2 import hash_password, verify_password
+from app.OAuth2 import hash_password
 from app.enums.emailTemplate import EmailTemplate
 from app.services import auth
 from app.utilities import div_ciel, send_mail
-from app.enums import AccountStatus, TokenStatus, Role
-from fastapi import HTTPException, status
-from .error import get_error_detail, add_error
-from datetime import datetime
+from app.enums import AccountStatus, Role
 from app.dependencies import PaginationParams
-
-error_keys = {
-    "ck_employees_cnss_number": {
-        "message": "-Cnss number it should be {8 digits}-{2 digits} and it's mandatory if contract type Cdi or Cdd",
-        "status": 400,
-    },
-    "employees_email_key": {"message": "Email already used", "status": 409},
-    "employees_number_key": {"message": "Number already used", "status": 409},
-    "Username and Password not accepted": {
-        "message": "Email credentials not accepted",
-        "status": 406,
-    },
-    "employees_pkey": {"message": "Employee not found", "status": 404},
-    "unique_employee_role": {
-        "message": "This employee already has this role",
-        "status": 409,
-    },
-}
 
 
 def convert_employee_to_schema(employee: models.Employee):
@@ -50,63 +28,47 @@ def convert_employee_to_schema(employee: models.Employee):
         created_at=employee.created_at,
     )
 
+def get_all(db: Session, pg_params: PaginationParams):
+    skip = pg_params.limit * (pg_params.page - 1)
+    query = db.query(models.Employee)
+    if pg_params.name:
+        query = query.filter(func.lower(func.concat(models.Employee.first_name, " ", models.Employee.last_name)).contains(func.lower(pg_params.name)))
+    total_records = query.count()
+    total_pages = div_ciel(total_records, pg_params.limit)
+    result = query.limit(pg_params.limit).offset(skip).all()
+    return schemas.EmployeesOut(
+            list=[convert_employee_to_schema(emp) for emp in result],
+            total_records=total_records,
+            total_pages=total_pages,
+            page_number=pg_params.page,
+            page_size=pg_params.limit,
+            detail="Employees fetched",
+            status_code=200,
+        )
 
 def update_roles(roles: List[Role], db: Session, employee_id: int):
     try:
-        query = db.query(models.EmployeeRole.role).filter(
-            models.EmployeeRole.employee_id == employee_id
-        )
+        query = db.query(models.EmployeeRole.role).filter(models.EmployeeRole.employee_id == employee_id)
         current_roles = {role[0] for role in query.all()}
         new_roles = {role for role in roles}
         roles_to_delete = current_roles - new_roles
         roles_to_add = new_roles - current_roles
         if roles_to_delete:
-            db.query(models.EmployeeRole).filter(
-                models.EmployeeRole.employee_id == employee_id,
-                models.EmployeeRole.role.in_(roles_to_delete),
-            ).delete()
+            db.query(models.EmployeeRole).filter(models.EmployeeRole.employee_id == employee_id,models.EmployeeRole.role.in_(roles_to_delete),).delete()
         if roles_to_add:
-            db.add_all(
-                [
-                    models.EmployeeRole(role=role, employee_id=employee_id)
-                    for role in roles_to_add
-                ]
-            )
+            db.add_all([models.EmployeeRole(role=role, employee_id=employee_id)for role in roles_to_add])
             return True
     except Exception as error:
         return schemas.BaseOut(status_code=400, detail=str(error))
 
-
-def get_all(db: Session, pg_params: PaginationParams):
-    skip = pg_params.limit * (pg_params.page - 1)
-    query = db.query(models.Employee)
-    if pg_params.name:
-        query = query.filter(
-            func.lower(
-                func.concat(models.Employee.first_name, " ", models.Employee.last_name)
-            ).contains(func.lower(pg_params.name))
-        )
-    total_records = query.count()
-    total_pages = div_ciel(total_records, pg_params.limit)
-    result = query.limit(pg_params.limit).offset(skip).all()
-    return {
-        "total_records": total_records,
-        "total_pages": total_pages,
-        "list": result,
-    }
-
-
 def get_by_id(id: int, db: Session):
     return db.query(models.Employee).filter(models.Employee.id == id).first()
-
 
 def sudo_edit_employee(id: int, new_data: dict, db: Session):
     return db.query(models.Employee).filter(models.Employee.id == id).update(new_data)
 
-
 def get_by_email(email: str, db: Session):
     return db.query(models.Employee).filter(models.Employee.email == email).first()
-
 
 async def add(employee_dict: dict, db: Session):
     roles = employee_dict.pop("roles")
@@ -126,11 +88,10 @@ async def add(employee_dict: dict, db: Session):
     db.commit()
     return new_emp
 
-
 async def update(employee_id: int, fields_to_update: dict, db: Session):
     employee_to_update = get_by_id(employee_id, db)
     if employee_to_update is None:
-        return schemas.BaseOut(status_code=status.HTTP_404_NOT_FOUND, detail="Employee not found")
+        return schemas.BaseOut(status_code=404, detail="Employee not found")
     {fields_to_update.pop(key, None) for key in ["confirm_password", "actual_password"]}
     if fields_to_update["password"] != None:
         fields_to_update["password"] = hash_password(fields_to_update["password"])
